+++ web/src/main/java/ua/com/vertex/controllers/ViewAllCoursesController.java
@@ -7,11 +7,11 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 import ua.com.vertex.beans.Course;
-import ua.com.vertex.beans.User;
-import ua.com.vertex.logic.interfaces.UserLogic;
+import ua.com.vertex.logic.interfaces.CourseLogic;
 import ua.com.vertex.utils.DataNavigator;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping(value = "/viewAllCourses")
@@ -20,15 +20,15 @@ public class ViewAllCoursesController {
     private static final String ERROR_JSP = "error";
     private static final String PAGE_JSP = "viewAllCourses";
 
-    private static final Logger LOGGER = LogManager.getLogger(UserController.class);
+    private static final Logger LOGGER = LogManager.getLogger(ViewAllCoursesController.class);
 
-    private UserLogic userLogic;
+    private CourseLogic courseLogic;
 
     @GetMapping
     public ModelAndView viewAllCourses(@ModelAttribute DataNavigator dataNavigator) {
         ModelAndView modelAndView = new ModelAndView();
         try {
-            List<Course> courses = userLogic.getCoursesPerPages(dataNavigator);
+            List<Course> courses = courseLogic.getCoursesPerPages(dataNavigator);
             modelAndView.addObject("viewAllCourses", dataNavigator);
             modelAndView.addObject("courses", courses);
             modelAndView.setViewName(PAGE_JSP);
@@ -50,15 +50,17 @@ public class ViewAllCoursesController {
     public ModelAndView reload(@ModelAttribute DataNavigator dataNavigator) {
         ModelAndView modelAndView = new ModelAndView();
         try {
-            List<Course> courses = userLogic.getCoursesPerPages(dataNavigator);
+            List<Course> courses = courseLogic.getCoursesPerPages(dataNavigator);
             modelAndView.addObject("courses", courses);
             modelAndView.addObject("viewAllCourses", dataNavigator);
             modelAndView.setViewName(PAGE_JSP);
             LOGGER.debug("Received a list of all courses and transferred to the model");
-            String allCourses = "";
-            for (Course course : courses) {
-                allCourses += course.getName() + "|";
-            }
+//            String allCourses = "";
+//            for (Course course : courses) {
+//                allCourses += course.getName() + "|";
+//            }
+
+            String allCourses = courses.stream().map(Course::getName).collect(Collectors.joining("|"));
             LOGGER.debug("Quantity courses -" + courses.size());
             LOGGER.debug("All courses list -" + allCourses);
         } catch (Exception e) {
@@ -74,8 +76,8 @@ public class ViewAllCoursesController {
     }
 
     @Autowired
-    public ViewAllCoursesController(UserLogic userLogic) {
-        this.userLogic = userLogic;
+    public ViewAllCoursesController(CourseLogic courseLogic) {
+        this.courseLogic = courseLogic;
     }
 }
