+++ web/src/main/java/ua/com/vertex/logic/UserLogicImpl.java
@@ -1,11 +1,9 @@
 package ua.com.vertex.logic;
 
-
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ua.com.vertex.beans.Course;
 import ua.com.vertex.beans.Role;
 import ua.com.vertex.beans.User;
 import ua.com.vertex.dao.interfaces.UserDaoInf;
@@ -76,31 +74,31 @@ public class UserLogicImpl implements UserLogic {
         return users;
     }
 
-    @Override
-    public List<Course> getCoursesPerPages(DataNavigator dataNavigator) {
-        List<Course> courses = null;
-        if (dataNavigator.getCurrentNumberPage() == dataNavigator.getNextPage()) {
-            dataNavigator.setCurrentNumberPage(1);
-            dataNavigator.setNextPage(1);
-        } else {
-            dataNavigator.setCurrentNumberPage(dataNavigator.getNextPage());
-        }
-
-        try {
-            courses = userDao.getAllCourses(dataNavigator);
-            int dataSize = userDao.getQuantityCourses();
-
-            dataNavigator.setDataSize(dataSize);
-            dataNavigator.setQuantityPages(dataSize / dataNavigator.getCurrentRowPerPage());
-            if (dataSize / dataNavigator.getCurrentRowPerPage() >= 0) {
-                dataNavigator.setQuantityPages(dataNavigator.getQuantityPages() + 1);
-            }
-            dataNavigator.setLastPage(dataNavigator.getQuantityPages());
-        } catch (SQLException e) {
-            LOGGER.warn(e);
-        }
-        return courses;
-    }
+//    @Override
+//    public List<Course> getCoursesPerPages(DataNavigator dataNavigator) {
+//        List<Course> courses = null;
+//        if (dataNavigator.getCurrentNumberPage() == dataNavigator.getNextPage()) {
+//            dataNavigator.setCurrentNumberPage(1);
+//            dataNavigator.setNextPage(1);
+//        } else {
+//            dataNavigator.setCurrentNumberPage(dataNavigator.getNextPage());
+//        }
+//
+//        try {
+//            courses = userDao.getAllCourses(dataNavigator);
+//            int dataSize = userDao.getQuantityCourses();
+//
+//            dataNavigator.setDataSize(dataSize);
+//            dataNavigator.setQuantityPages(dataSize / dataNavigator.getCurrentRowPerPage());
+//            if (dataSize / dataNavigator.getCurrentRowPerPage() >= 0) {
+//                dataNavigator.setQuantityPages(dataNavigator.getQuantityPages() + 1);
+//            }
+//            dataNavigator.setLastPage(dataNavigator.getQuantityPages());
+//        } catch (SQLException e) {
+//            LOGGER.warn(e);
+//        }
+//        return courses;
+//    }
 
     @Override
     public Optional<User> getUserDetailsByID(int userId) throws SQLException {
