+++ web/src/test/java/ua/com/vertex/controllers/ViewAllCoursesControllerTest.java
@@ -13,6 +13,7 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.servlet.view.InternalResourceView;
 import ua.com.vertex.context.TestConfig;
+import ua.com.vertex.logic.interfaces.CourseLogic;
 import ua.com.vertex.logic.interfaces.UserLogic;
 
 import static org.hamcrest.Matchers.*;
@@ -29,7 +30,7 @@ import java.util.TreeMap;
 @ActiveProfiles("test")
 public class ViewAllCoursesControllerTest {
     @Autowired
-    private UserLogic logic;
+    private CourseLogic logic;
 
     private MockMvc mockMvc;
